WR457626: Prune historic trend data
A new scheduled task that runs hourly to prune historic trend data.
Long-running assessments may generate a large number of trend data
points. We have a trendlimit setting to control how many points are
plotted, but additional points currently remain in the database. This
task allows points that will not be plotted to be periodically cleaned
up, reducing the table size and speeding up queries.