WR457639: Filter attempts by current quiz
The query to report on quiz attempts was joining on the user's latest
attempt for any quiz, rather than the current quiz. The resulted in the
report showing attempt statuses for users that did not match their
actual attempt statuses when viewing the quiz's own report.

This fix adds the quiz ID as a parameter to the subquery, to ensure we
only find the latest attempt for the current quiz.